backend/matrimony/requirement.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.2.13
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
executing==2.0.1
ipython==8.12.3
ipython-genutils==0.2.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.1
wcwidth==0.2.13
